Refactor test files and update imports for DeleteButton and Metadata components
 - Move DeleteButton.test.tsx and Metadata.test.tsx from src/components to __tests__ directory.
 - Update import paths to reflect the new location of the test files.
 - Replace the mock implementation of useRouter from 'next/router' with 'next/navigation' in DeleteButton.test.tsx.
 - Use mockRouter from "next-router-mock" in DeleteButton.test.tsx.
 - Remove unused imports such as React and '@testing-library/jest-dom' from Metadata.test.tsx.
 - Adjust imports in both test files to import components from their new relative paths.
